refactor: remove component file to PaymentSystem folder
