+++ src/components/PaymentSystem/MovieInfoCard.js
