+++ src/pages/payment/index.js
@@ -1,4 +1,4 @@
-import PaymentMobile from '../../components/PaymentMobile';
+import PaymentMobile from '../../components/PaymentSystem/PaymentMobile';
 
 export default function PaymentPage() {
   return <PaymentMobile />;
